services/query-analyzer/requirements.txt
# Core service dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and HTTP
httpx==0.25.2
asyncio-mqtt==0.16.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration
PyYAML==6.0.1

# gRPC support (for future Phase 1.3)
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# ML and NLP dependencies (from existing Epic1 components)
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.6.0

# Additional dependencies that Epic1QueryAnalyzer might need
pandas>=2.0.0
scipy>=1.10.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # for testing

# Optional: if we need database connections later
# psycopg2-binary==2.9.7
# redis==5.0.1